requirements.txt
# ============================================================
# Core LangGraph / LangChain Stack
# ============================================================
langgraph>=0.2.28
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langsmith>=0.1.120

# ============================================================
# Azure OpenAI
# ============================================================
openai>=1.40.0

# ============================================================
# Pydantic (v2) for schema validation
# ============================================================
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ============================================================
# Scheduler
# ============================================================
APScheduler>=3.10.4

# ============================================================
# Databricks SDK (production-like interface)
# ============================================================
databricks-sdk>=0.28.0

# ============================================================
# FastMCP (MCP tooling layer)
# ============================================================
fastmcp>=0.4.0

# ============================================================
# Email
# ============================================================
sendgrid>=6.11.0

# ============================================================
# Observability - Arize Phoenix (local OSS)
# ============================================================
arize-phoenix>=4.0.0
openinference-instrumentation-langchain>=0.1.20
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0

# ============================================================
# Persistence
# ============================================================
aiosqlite>=0.20.0
SQLAlchemy>=2.0.30

# ============================================================
# HTTP / Async
# ============================================================
httpx>=0.27.0
aiohttp>=3.9.5
tenacity>=8.3.0

# ============================================================
# Utilities
# ============================================================
python-dotenv>=1.0.1
structlog>=24.2.0
rich>=13.7.1
uuid6>=2024.1.12
python-dateutil>=2.9.0

# ============================================================
# Testing
# ============================================================
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-mock>=3.14.0
